requirements.txt
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.2
nltk>=3.7
streamlit>=1.10.0
gtts>=2.2.4
googletrans==4.0.0-rc1
Pillow>=9.0.0